Address PR feedback - remove some code duplication.
